% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{configs_job}
\alias{configs_job}
\title{get configs of a job}
\usage{
configs_job(project_id, job_id, folder)
}
\arguments{
\item{project_id}{project id of the job corresponding project}

\item{job_id}{job id of the job of interest}

\item{folder}{folder name, where to save the resulting files}
}
\value{
list of filenames of the resulting files
}
\description{
this function downloads the configs of a job into a defined folder under the folder name projects/project_id/jobs/job_id
}
