% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{create_job}
\alias{create_job}
\title{create a job}
\usage{
create_job(project_id, params = list())
}
\arguments{
\item{project_id}{project_id of where to create a new job}
}
\value{
job id of created job
}
\description{
Given a project_id a new job is created. The model and revision of the model is defined by the project.
}
