% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{create_project}
\alias{create_project}
\title{create project}
\usage{
create_project(model_id, revision = "HEAD", params = list())
}
\arguments{
\item{model_id}{defines the model to start the project with}

\item{params}{defines the parameter different from the
standard parameter of the first job; it is recommended to call
get_params first and then to modify the list and put here as argument.}
}
\value{
id of the created project, error otherwise
}
\description{
create project
}
