% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{delete_job}
\alias{delete_job}
\title{deletes a job}
\usage{
delete_job(project_id, job_id)
}
\arguments{
\item{project_id}{project id of job to be deleted}

\item{job_id}{job id of the job to be deleted}
}
\value{
the list of remaining jobs within the given project
}
\description{
deletes a job
}
