% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{delete_project}
\alias{delete_project}
\title{deletes a job according to the given id}
\usage{
delete_project(project_id = NA)
}
\arguments{
\item{project_id}{defines the project to be deleted}
}
\value{
if success: remaining projects, the error otherwise
}
\description{
deletes a job according to the given id
}
