% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{get_auth_token}
\alias{get_auth_token}
\title{get function for the authenticity token}
\usage{
get_auth_token()
}
\value{
the authenticity token, if logged in, NA otherwise
}
\description{
get function for the authenticity token
}
