% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{get_jobs_status}
\alias{get_jobs_status}
\title{get jobs of a given project}
\usage{
get_jobs_status(project_id)
}
\arguments{
\item{project_id}{corresponding project id of the jobs looking for}
}
\value{
data.frame of ids and status of the jobs within the given project
}
\description{
get jobs of a given project
}
