% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_params}
\alias{get_params}
\title{parameter of a given model}
\usage{
get_params(model_id)
}
\arguments{
\item{model_id}{the given model based on the id}
}
\value{
a list of parameter
}
\description{
NOTE: no file upload support yet
}
