% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{get_projects}
\alias{get_projects}
\title{gives back all existing projects of the logged in user}
\usage{
get_projects()
}
\value{
list of all existing projects
}
\description{
gives back all existing projects of the logged in user
}
