% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Logs into a magpie environment}
\usage{
login(email, password)
}
\arguments{
\item{email}{email of the login form}

\item{password}{password of the login form}
}
\value{
Success if login was succesfull, the error otherwise.
}
\description{
Logs into a magpie environment
}
