% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAll.R
\name{PlotAll}
\alias{PlotAll}
\title{All plots of power evaulation results under various scenarios.}
\usage{
PlotAll(Power.list, depth_factor = 1)
}
\arguments{
\item{Power.list}{A list produced by \code{\link{PowerEval}}.}

\item{depth_factor}{A numerical value indicating which sequencing depth to plot. For example, 2 means doubling the original sequencing depth. Default is 1.}
}
\value{
It plots all power measurements of a certain sequencing depth in a 2x2 panel. Power measurements to plot include "FDR", "FDC", "Power", and "Precision".
}
\description{
This function plots all power measurements of a certain sequencing depth in a 2x2 panel. Power measurements to plot include "FDR", "FDC", "Power", and "Precision".
}
\examples{
\donttest{
### Main function
power.test <- PowerEval(
    Input.file = c(
        "Ctrl1.chr1.input.bam", "Ctrl2.chr1.input.bam",
        "Case1.chr1.input.bam", "Case2.chr1.input.bam"
    ),
    IP.file = c(
        "Ctrl1.chr1.ip.bam", "Ctrl2.chr1.ip.bam",
        "Case1.chr1.ip.bam", "Case2.chr1.ip.bam"
    ),
    BamDir = "./data/GSE46705_split_chr",
    annoDir = "./data/annotation/hg18_chr1.sqlite",
    variable = rep(c("Ctrl", "Trt"), each = 2),
    bam_factor = 0.08,
    nsim = 10,
    N.reps = c(2, 3, 5, 7),
    depth_factor = c(1, 2, 5),
    thres = c(0.01, 0.05, 0.1),
    Test_method = "TRESS"
)
### plot all figures
PlotAll(power.test, depth_factor = 2)
}
}
