% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRes.R
\name{PlotRes}
\alias{PlotRes}
\title{An individual Plot of power evaulation results under various scenarios.}
\usage{
PlotRes(Power.list, depth_factor = 1, value_option = "FDR")
}
\arguments{
\item{Power.list}{A list produced by \code{\link{PowerEval}}.}

\item{depth_factor}{A numerical value indicating which sequencing depth to plot. For example, 2 means doubling the original sequencing depth. Default is 1.}

\item{value_option}{A character indicating which measurement to plot. Options include "FDR", "FDC", "Power", and "Precision".}
}
\value{
It plots a certain power measurement of a certain sequencing depth. Power measurements to plot include "FDR", "FDC", "Power", and "Precision".
}
\description{
This function plots a certain power measurement of a certain sequencing depth. Power measurements to plot include "FDR", "FDC", "Power", and "Precision".
}
\examples{
\donttest{
### Main function
power.test <- PowerEval(
    Input.file = c(
        "Ctrl1.chr1.input.bam", "Ctrl2.chr1.input.bam",
        "Case1.chr1.input.bam", "Case2.chr1.input.bam"
    ),
    IP.file = c(
        "Ctrl1.chr1.ip.bam", "Ctrl2.chr1.ip.bam",
        "Case1.chr1.ip.bam", "Case2.chr1.ip.bam"
    ),
    BamDir = "./data/GSE46705_split_chr",
    annoDir = "./data/annotation/hg18_chr1.sqlite",
    variable = rep(c("Ctrl", "Trt"), each = 2),
    bam_factor = 0.08,
    nsim = 10,
    N.reps = c(2, 3, 5, 7),
    depth_factor = c(1, 2, 5),
    thres = c(0.01, 0.05, 0.1),
    Test_method = "TRESS"
)
### plot an individual figure
PlotRes(power.test, depth_factor = 5, value_option = "FDR")
}
}
