% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE55575_TRESS_res.R
\docType{data}
\name{power.test}
\alias{power.test}
\title{Power calculation results for GSE55575}
\format{
A list object.
}
\usage{
data(GSE55575_TRESS_res)
}
\description{
Mouse embryonic fibroblasts: Two replicates of wild type (WT)
and four replicates of knockdown (KD) of WTAP.
}
\examples{
data(GSE55575_TRESS_res)
}
\keyword{datasets}
