% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{reset_url}
\alias{reset_url}
\title{Deletes the manually curated url in .magpie_data}
\usage{
reset_url()
}
\value{
the actual configured url, used for all magpie api operations.
}
\description{
When calling this function, the preconfigured server address to the demo server
at https://magpie.imb.medizin.tu-dresden.de is used.
}
