% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{set_url}
\alias{set_url}
\title{Sets the url used for all magpie api operations}
\usage{
set_url(url = "https://magpie.imb.medizin.tu-dresden.de")
}
\arguments{
\item{url}{url used for all following api operations}
}
\value{
the actual configured url, used for all magpie api operations; can differ if protocoll is missing.
}
\description{
Sets the actual url used for all api operations in global environment. It is stored in magpie_url.
If url is not set manually, the demo server at https://magpie.imb.medizin.tu-dresden.de is autmatically
chosen for all api operations.
}
