% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteToxlsx.R
\name{WriteToxlsx}
\alias{WriteToxlsx}
\title{Write power evaulation results under all scenarios to a .xlsx file.}
\usage{
WriteToxlsx(pl, file)
}
\arguments{
\item{pl}{A list produced by \code{\link{PowerEval}}.}

\item{file}{A character indicating the name of the output .xlsx file.}
}
\value{
It outputs a .xlsx file including FDR, FDC, power, and precision under various sample sizes,
sequencing depths, and adjusted p-value thresholds.
}
\description{
This function writes power evaulation results to a .xlsx file.
}
\examples{

library(PAM)
### Main function
power.test <- QuickPower(dataset = "GSE55575", test_method = "TRESS")

### write out .xlsx
WriteToxlsx(power.test, file = "test_TRESS.xlsx")

}
