% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteToxlsx_strata.R
\name{WriteToxlsx_strata}
\alias{WriteToxlsx_strata}
\title{Write power evaulation results of four strata under all sample size scenarios to a .xlsx file.}
\usage{
WriteToxlsx_strata(pl, file)
}
\arguments{
\item{pl}{A list produced by \code{\link{PowerEval}}.}

\item{file}{A character indicating the name of the output .xlsx file.}
}
\value{
It outputs a .xlsx file including FDR, FDC, power, and precision under the original sequencing depth and various sample sizes
and input stratas.
}
\description{
This function writes power evaulation results of four strata to a .xlsx file. Only results from the original sequencing depth are saved.
    Here, strata are determined by mean input control levels of simulated data.
}
\examples{

library(PAM)
### Main function
power.test <- QuickPower(dataset = "GSE55575", test_method = "TRESS")

### write out .xlsx
WriteToxlsx_strata(power.test, file = "test_strata_TRESS.xlsx")

}
