% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotALL_Strata.R
\name{PlotALL_Strata}
\alias{PlotALL_Strata}
\title{All plots of power evaulation results by strata.}
\usage{
PlotALL_Strata(Power.list)
}
\arguments{
\item{Power.list}{A list produced by \code{\link{PowerEval}}.}
}
\value{
It plots all power measurements of the original sequencing depth by strata in a 2x2 panel. Power measurements to plot include "FDR", "FDC", "Power", and "Precision".
}
\description{
This function plots all power measurements of the original sequencing depth by strata in a 2x2 panel. Power measurements to plot include "FDR", "FDC", "Power", and "Precision".
}
\examples{

library(PAM)
### Main function
power.test <- QuickPower(dataset = "GSE55575", test_method = "TRESS")

### plot all strata results in a panel
PlotALL_Strata(power.test)

}
