% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotStrata.R
\name{PlotStrata}
\alias{PlotStrata}
\title{An individual plot of power evaulation results by strata.}
\usage{
PlotStrata(Power.list, value_option = "FDR")
}
\arguments{
\item{Power.list}{A list produced by \code{\link{PowerEval}}.}

\item{value_option}{A character indicating which measurement to plot. Options include "FDR", "FDC", "Power", and "Precision".}
}
\value{
It plots a certain power measurement of the original sequencing depth by strata. Power measurements to plot include "FDR", "FDC", "Power", and "Precision".
}
\description{
This function plots a certain power measurement of the original sequencing depth by strata. Power measurements to plot include "FDR", "FDC", "Power", and "Precision".
}
\examples{

library(PAM)
### Main function
power.test <- QuickPower(dataset = "GSE55575", test_method = "TRESS")

### plot a FDR strata result
PlotStrata(power.test, value_option = "FDR")

}
