% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSE46705_TRESS_res.R
\docType{data}
\name{power.test}
\alias{power.test}
\title{Power calculation results for GSE46705}
\format{
A list object.
}
\usage{
data(GSE46705_TRESS_res)
}
\description{
Human HeLa cell line:
Two replicates of wild type (WT) and two replicates of knockdown (KD) of complex METTL3.
}
\examples{
data(GSE46705_TRESS_res)
}
\keyword{datasets}
