% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteToxlsx_strata.R
\name{WriteToxlsx_strata}
\alias{WriteToxlsx_strata}
\title{Write power evaulation results of four strata under all sample size scenarios to a .xlsx file.}
\usage{
WriteToxlsx_strata(pl, file)
}
\arguments{
\item{pl}{A list produced by \code{\link{PowerEval}}.}

\item{file}{A character indicating the name of the output .xlsx file.}
}
\value{
It outputs a .xlsx file including FDR, FDC, power, and precision under the original sequencing depth and various sample sizes
and input stratas.
}
\description{
This function writes power evaulation results of four strata to a .xlsx file. Only results from the original sequencing depth are saved.
    Here, strata are determined by mean input control levels of simulated data.
}
\examples{
\donttest{
### Main function
power.test <- PowerEval(
    Input.file = c(
        "Ctrl1.chr1.input.bam", "Ctrl2.chr1.input.bam",
        "Case1.chr1.input.bam", "Case2.chr1.input.bam"
    ),
    IP.file = c(
        "Ctrl1.chr1.ip.bam", "Ctrl2.chr1.ip.bam",
        "Case1.chr1.ip.bam", "Case2.chr1.ip.bam"
    ),
    BamDir = "./data/GSE46705_split_chr",
    annoDir = "./data/annotation/hg18_chr1.sqlite",
    variable = rep(c("Ctrl", "Trt"), each = 2),
    bam_factor = 0.08,
    nsim = 10,
    N.reps = c(2, 3, 5, 7),
    depth_factor = c(1, 2, 5),
    thres = c(0.01, 0.05, 0.1),
    Test_method = "TRESS"
)
### write strata results
WriteToxlsx_strata(power.test, file = "test1_:strata.xlsx")
}
}
